Create index.html
Q.	How did God reveal His words to man?
A.	Spoken by God  	Heb. 1:1-2

Q.	How were the words of God recorded and preserved until today? How were these words handed down to us?
A.	He commanded His servants to write  	  Jer. 30:2

Q.	Which book contains the words of God, in its entirety?
A.	The Holy Scriptures          .......................................   II Tim. 3:15
	
Q. 	What makes the Bible unique?  Why do we regard it as above all books?
A.	The Bible is inspired by God  ..............................   II Tim. 3:16-17

Q.	How were those commissioned to write the Bible prevented from committing errors?   Since it is just human nature to err (is human)?
A.	The writers were guided 	  Rev. 10.4

Q.	What is one of the characteristics of the words of God?
A.	Truth  	 Jn. 17:17

Q.	What does it mean?  Why is it said to be the truth?
A.	Because God’s word have fulfillment 	  Is. 46:11

Q.	What did God say through our Lord Jesus Christ in the beginning of the Christian era that He said would happen near the end of the Christian era?
A.	The two world wars  	Mt. 24:6-7

Q.	What else did God say through our Lord Jesus Christ in the beginning of the Christian era but destined to take place at some later time?
A.	followed by famine and poverty  	Mt. 24:7-8

Q.	What did God say through the Prophet Daniel concerning the knowledge of man during the last days?
A.	Knowledge shall increase  	 Dan. 12:4

Q.	What about the kind of people that will live during these days?  What did God say by means of the Apostles about the kind of people that will live during these days?
A.	Perilous times  ........................ ……………………….   II Tim. 3:1-5

Q.	What is the value of believing in the words of God written in the Bible?
A.	May have life everlasting ........................................  Jn. 20:30-31

Q.	Thus, what is the strict instruction of the Apostles so that we may not fall prey to erroneous faiths?
A.	Do not go beyond what is written 	 I Cor. 4:6
	
Q.	What is God’s warning to those who add to or subtract from the words of God written in the Bible?  What is wrong with that?
A.	They will be punished and will not be saved  ……………….  Rev. 22:18-19